requirements.txt
tensorflow>=2.13.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
librosa>=0.10.0
PyYAML>=6.0
tqdm>=4.65.0
pillow>=10.0.0
pyaudio>=0.2.11
pygame>=2.5.0